+++ Engine/SSTextureCube.cpp
@@ -33,19 +33,19 @@ bool SSTextureCube::LoadFromDDSFile(std::wstring filename)
 	}
 	
 	mWidth = static_cast<UINT>(metaData.width);
-	mHeight = metaData.height;
-	mMipLevels = metaData.mipLevels;
+	mHeight = static_cast<UINT>(metaData.height);
+	mMipLevels = static_cast<UINT>(metaData.mipLevels);
 
 	D3D11_TEXTURE2D_DESC description;
-	description.Width = mWidth = metaData.width;
-	description.Height = mHeight = metaData.height;
+	description.Width = mWidth = static_cast<UINT>(metaData.width);
+	description.Height = mHeight = static_cast<UINT>(metaData.height);
 	description.BindFlags = D3D11_BIND_SHADER_RESOURCE;
 	description.MiscFlags = D3D11_RESOURCE_MISC_TEXTURECUBE;
 	description.Usage = D3D11_USAGE_DEFAULT;
 	description.SampleDesc.Count = 1;
 	description.SampleDesc.Quality = 0;
-	description.MipLevels = metaData.mipLevels;
-	description.ArraySize = metaData.arraySize;
+	description.MipLevels = static_cast<UINT>(metaData.mipLevels);
+	description.ArraySize = static_cast<UINT>(metaData.arraySize);
 	description.CPUAccessFlags = 0;
 	
 	if (bSRGB)
@@ -64,7 +64,7 @@ bool SSTextureCube::LoadFromDDSFile(std::wstring filename)
 	resourceViewDesc.Format = description.Format;
 	resourceViewDesc.ViewDimension = D3D11_SRV_DIMENSION_TEXTURECUBE;
 	resourceViewDesc.Texture2D.MostDetailedMip = 0;
-	resourceViewDesc.Texture2D.MipLevels = metaData.mipLevels;
+	resourceViewDesc.Texture2D.MipLevels = static_cast<UINT>(metaData.mipLevels);
 
 	HR(SSEngine::Get().GetDevice()->CreateShaderResourceView(mTexturePtr, &resourceViewDesc, &mShaderResourceView));
